requirements.txt
# Core
pytz==2018.9
Django==2.2
gunicorn==19.9.0
newrelic==4.12.0.113
django-dotenv==1.4.2

# For the persistence stores
mysqlclient==1.4.4
redis==3.3.11

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1
django-summernote==0.8.11.5
django-celery-beat==1.3.0

# Rest apis
djangorestframework==3.11.0
Markdown==3.0.1
drf-yasg==1.17.0
django-filter==2.1.0
django-cors-headers==3.2.0
django-rest-passwordreset==1.0.0
django-rest-swagger==2.2.0
django-versatileimagefield==1.10
django-auto-prefetching==0.1.8

# Social login
social-auth-core==3.2.0
social-auth-app-django==3.1.0

# Developer Tools
ipdb==0.11
ipython==7.2.0
flake8==3.7.5
yapf==0.28.0
sentry-sdk==0.11.2
django-inlinecss==0.3.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.2
coverage==4.5.2

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.93

